+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
  - diagnostic for detecting unnecessary script imports when `autoImportComponentScript` is enabled
 ### Fixed
+ - bug that was flagging namespaced functions with the same name as a stdlib function.
  - bug that was not properly transpiling brighterscript script tags in xml components.
  - several performance issues introduced in v0.8.2. 
  - Replace `type="text/brighterscript"` with `type="text/brightscript"` in xml script imports during transpile. ([#73](https://github.com/rokucommunity/brighterscript/issues/73))